# src/swagger.yaml
swagger: '2.0'
info:
  description: |
    This is a movie search engine app with all the CRUD REST API's with local database on mongodb.
  version: 1.0.0
  title: Movie Search Engine App
  termsOfService: https://github.com/Krishan27/movie_api
  contact:
    email: krishansharma0327@gmail.com
  license:
    name: MIT
    url: https://github.com/Krishan27/movie_api
# host: http://localhost:5000
# basePath: /api
...
paths:
  /movies/omdb:
    get:
      tags:
      - movie from omdb
      summary: Search the movie for the user in omdb database.
      operationId: getmovie
      consumes:
      - application/json
      
      produces:
      - application/json
      
      parameters:
      - in: searchId
        name: searchName
        description: search movie in omdb database by movieId or movieName
        required: true
        schema:
          $ref: './controllers/movieController.js'
      responses:
        404:
          description: movie not found
        500:
          description: Failed
      
    get:
      tags:
      - movie from 
      summary: Update an existing movie
      operationId: updatePet
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: movie object that needs to be added to the store
        required: true
        schema:
          $ref: '#/definitions/movie'
          
      responses:
        400:
          description: Invalid ID supplied
        404:
          description: movie not found
        405:
          description: Validation exception
      security:
      - petstore_auth:
        - write:movies
        - read:movies
...
  /movie/{petId}:
    get:
      tags:
      - movie
      summary: Find movie by ID
      description: Returns a single movie
      operationId: getPetById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: petId
        in: path
        description: ID of movie to return
        required: true
        type: integer
        format: int64
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/movie'
        400:
          description: Invalid ID supplied
        404:
          description: movie not found
      security:
      - api_key: []
    post:
      tags:
      - movie
      summary: Updates a movie in the store with form data
      operationId: updatePetWithForm
      consumes:
      - application/x-www-form-urlencoded
      produces:
      - application/json
      - application/xml
      parameters:
      - name: petId
        in: path
        description: ID of movie that needs to be updated
        required: true
        type: integer
        format: int64
      - name: name
        in: formData
        description: Updated name of the movie
        required: false
        type: string
      - name: status
        in: formData
        description: Updated status of the movie
        required: false
        type: string
      responses:
        405:
          description: Invalid input
      security:
      - petstore_auth:
        - write:movies
        - read:movies
    delete:
      tags:
      - movie
      summary: Deletes a movie
      operationId: deletePet
      produces:
      - application/json
      - application/xml
      parameters:
      - name: api_key
        in: header
        required: false
        type: string
      - name: petId
        in: path
        description: movie id to delete
        required: true
        type: integer
        format: int64
      responses:
        400:
          description: Invalid ID supplied
        404:
          description: movie not found
      security:
      - petstore_auth:
        - write:movies
        - read:movies
...
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: http://petstore.swagger.io/oauth/dialog
    flow: implicit
    scopes:
      write:movies: modify movies in your account
      read:movies: read your movies
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
...
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Category
...
  Tag:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Tag
  movie:
    type: object
    required:
    - name
    - photoUrls
    properties:
      id:
        type: integer
        format: int64
      category:
        $ref: '#/definitions/Category'
      name:
        type: string
        example: doggie
      photoUrls:
        type: array
        xml:
          name: photoUrl
          wrapped: true
        items:
          type: string
      tags:
        type: array
        xml:
          name: tag
          wrapped: true
        items:
          $ref: '#/definitions/Tag'
      status:
        type: string
        description: movie status in the store
        enum:
        - available
        - pending
        - sold
    xml:
      name: movie
...
host: localhost:5000
basePath: /api
schemes:
 - https
 - http